mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.49.0",
    "bs4>=0.0.2",
    "langchain>=0.3.21",
    "langchain-mcp-adapters>=0.0.5",
    "langchain-openai>=0.3.9",
    "langgraph>=0.3.18",
    "mcp>=1.4.1",
    "python-dotenv>=1.0.1",
]
